tokenize_into_words now in another file, so it can also be used by the embedding.
