+++ shift_detector/precalculations/TextMetadata.py
@@ -15,6 +15,7 @@ from spellchecker import SpellChecker
 from textstat import textstat
 
 from shift_detector.precalculations.Precalculation import Precalculation
+from shift_detector.precalculations.Tokenizer import TokenizeIntoWords
 from shift_detector.utils import UCBlist
 from shift_detector.utils.ColumnManagement import ColumnType
 from shift_detector.utils.TextMetadataUtils import dictionary_to_sorted_string, delimiters
@@ -119,35 +120,6 @@ class GenericTextMetadataWithTokenizingAndLanguage(GenericTextMetadata):
             metadata1[column] = temp_column1
             metadata2[column] = temp_column2
         return metadata1, metadata2
-
-
-class TokenizeIntoWords(Precalculation):
-
-    def __eq__(self, other):
-        return isinstance(other, self.__class__)
-
-    def __hash__(self):
-        return hash(self.__class__)
-
-    def tokenize_into_words(self, text):
-        text = text.lower()
-        text = re.sub(r"-", ' ', text)
-        text = re.sub(r"[^\w\s']", '', text)
-        splitted = re.split(r'\W\s|\s', text)
-        while '' in splitted:
-            splitted.remove('')
-        return splitted
-
-    def process(self, store):
-        tokenized1 = pd.DataFrame()
-        tokenized2 = pd.DataFrame()
-        df1, df2 = store[ColumnType.text]
-        for column in df1.columns:
-            clean1 = df1[column].dropna()
-            clean2 = df2[column].dropna()
-            tokenized1[column] = [self.tokenize_into_words(text) for text in clean1]
-            tokenized2[column] = [self.tokenize_into_words(text) for text in clean2]
-        return tokenized1, tokenized2
 
 
 class NumCharsMetadata(GenericTextMetadata):
